+++ packages/terra-site/CHANGELOG.md
@@ -3,8 +3,9 @@ Changelog
 
 Unreleased
 ----------
-### Changed
+### Added
 * Add terra-form-textarea components
+* Add terra-form-field components
 * Add terra-mock-theme styles
 
 1.17.0 - (November 7, 2017)